requirements.txt
flask
flask-cors
stripe
itsdangerous
gunicorn
Pillow
reportlab
